SCREEN-131 完成图片列表的分页显示

+++ server-core/src/main/java/io/loli/sc/server/service/UploadedImageService.java
@@ -217,6 +217,6 @@ public class UploadedImageService {
     }
 
     public int countByGalIdAndUId(int id, Integer gid) {
-        return ud.countByGalIdAndUId(id, gid);
+        return ud.countByGalIdAndUId(id, gid).intValue();
     }
 }
